refactor(api): agents orchestrate·tmux 라우트를 전용 모듈로 추출
13개 라우트를 두 모듈로 나눴다 — orchestrate.py 6(태스크 분석·이력·실행),
tmux.py 7(Claude 인증 상태·tmux 세션 실행/조회/중단·실행 전략).
라우트 테이블 characterization 통과(29건 동일, 가림 쌍 0건).

**한 모듈로 두려다 800줄을 넘겨(862줄) 계획의 분기 지시대로 한 번 더 갈랐다.**
경계는 원본 섹션 배너를 그대로 따랐다.

workspace 경로 검증 3종(get_allowed_workspace_roots·_validate_project_path·
ALLOWED_WORKSPACE_ROOTS)은 tmux.py 로 갔다 — core·orchestrate 어느 쪽에서도
쓰이지 않고 오직 _validate_project_path 내부에서만 참조되는데, 그 함수는
tmux 실행 경로를 검증한다. __init__.py 의 재노출 출처를 .tmux 로 바꿨다.

이미지 업로드 상수 4종은 _shared.py 로 승격했다. ALLOWED_IMAGE_EXTENSIONS 와
MAX_IMAGE_SIZE 를 OCR(core)과 analyze-with-images(orchestrate) 양쪽이 쓰므로
어느 한쪽 소유로 둘 수 없다.

모든 파일이 800줄 한도 이내가 됐다: _legacy 583 · orchestrate 476 ·
tmux 412 · mcp 322 · _shared 17 · __init__ 47.

게이트: mypy 0, pytest 1348 passed(플레이크 1건 제외).
